# Skip splitmask calculations until we reach minSize