In my tests it gives ~1% perf improvement, though margins are quite narrow and likely newer versions of the compiler would be able optimize this out automagically.

```
$ benchstat old new                                       
name       old time/op   new time/op   delta
Chunker-8   58.2ms ± 3%   57.4ms ± 4%  -1.45%  (p=0.012 n=18+19)

name       old speed     new speed     delta
Chunker-8  576MB/s ± 3%  585MB/s ± 4%  +1.68%  (p=0.004 n=17+19)
```