@MichaelEischer 2022-05-07
This essentially reverts https://github.com/restic/chunker/commit/be7f5d823c60401be30ac77dcab9e824e4017691 , which improved performance for older CPUs by about 5%. The question where that check should be located unfortunately depends a lot on the branch prediction unit of the used CPU.